in dev

+++ Assignment2/opendlv-virtual-motor-kiwi/src/differential.cpp
@@ -75,7 +75,7 @@ opendlv::sim::KinematicState Differential::step(double dt) noexcept
     // calculations for yaw_rate vx vy from lecture notes
     float v = (this->vl + this->vr) / 2.0f;
     this->yaw_rate = (this->vr - this->vl) / (2*R);    // phi_dot
-    this->yaw = this->yaw + this->yaw_rate*dt;            // using the logic: phi(t+1) = phi(t) + phi_dot*dt
+    this->yaw = this->yaw + this->yaw_rate * static_cast<float>(dt);            // using the logic: phi(t+1) = phi(t) + phi_dot*dt
     this->vx = v*cos(this->yaw);
     this->vy = v*sin(this->yaw);
